Move all export-related code to ExportUtils
Fix #51
