+++ ropa/backend/backend.py
@@ -1,11 +1,5 @@
 from ropper import RopperService
 import json
-import struct
-import subprocess
-import sys
-
-# poppopret: pop ???; pop ???; ret
-# jmpreg: jmp ???
 
 
 class Backend:
@@ -141,84 +135,6 @@ class Backend:
                 gadgets[i]['info'] = 'pop-pop-ret'
 
         return gadgets
-
-    #######################################
-    # EXPORTATION TOOLS
-    #######################################
-
-    def num_bits(self, arch):
-        if arch.endswith('64'):
-            return 64
-        return 32
-
-    def open_exported(self, filepath):
-        if sys.platform.startswith('linux'):
-            subprocess.call(["xdg-open", filepath])
-
-        # Let's ignore this for now,
-        # I don't even know how to set python qt up on windows
-        #
-        # elif sys.platform.startswith('win'):
-        #     subprocess.call([filepath])
-
-        elif sys.platform.startswith('darwin'):  # mac
-            subprocess.call(["open", filepath])
-
-    def export_binary(self, file, chain):
-        with open(file, 'w') as outfile:
-            for block in chain:
-                for gadget in block:
-                    if self.num_bits(self.get_arch()) == 32:
-                        outfile.write(struct.pack('<I',
-                                      int(gadget['address'], 16)))
-                    else:
-                        outfile.write(struct.pack('<Q',
-                                      int(gadget['address'], 16)))
-            outfile.close()
-
-        self.open_exported(file)
-
-    def export_python_struct(self, file, chain):
-        with open(file, 'w') as outfile:
-            outfile.write('p = ""\n')
-            for block in chain:
-                for gadget in block:
-                    if self.num_bits(self.get_arch()) == 32:
-                        outfile.write('p += struct.pack("<I", {})'
-                                      .format(gadget['address']))
-                    else:
-                        outfile.write('p += struct.pack("<Q", {})'
-                                      .format(gadget['address']))
-
-                    outfile.write('  # ')
-                    for instruction in gadget['instructions']:
-                        outfile.write('{}; '.format(instruction))
-
-                    outfile.write('\n')
-            outfile.close()
-
-        self.open_exported(file)
-
-    def export_python_pwntools(self, file, chain):
-        with open(file, 'w') as outfile:
-            outfile.write('p = ""\n')
-            for block in chain:
-                for gadget in block:
-                    if self.num_bits(self.get_arch()) == 32:
-                        outfile.write('p += p32({})'
-                                      .format(gadget['address']))
-                    else:
-                        outfile.write('p += p64({})'
-                                      .format(gadget['address']))
-
-                    outfile.write('  # ')
-                    for instruction in gadget['instructions']:
-                        outfile.write('{}; '.format(instruction))
-
-                    outfile.write('\n')
-            outfile.close()
-
-        self.open_exported(file)
 
     #######################################
     # PROJECT TOOLS
