+++ ropa/gui/controller/__init__.py
@@ -1,6 +1,5 @@
 from file_dialog_controller import FileDialogController  # noqa
 from badbytes_input_controller import BadbytesInputController  # noqa
-from export_controller import ExportController  # noqa
 from filter_input_controller import FilterInputController  # noqa
 from list_key_controller import ListKeyController 	# noqa
 from list_widget_controller import ListWidgetController  # noqa
